% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateDensityZOIB.R
\name{bivariateDensityZOIB}
\alias{bivariateDensityZOIB}
\title{Bivariate Zero-One-Inflated Beta Density}
\usage{
bivariateDensityZOIB(
  x1,
  p1,
  q1,
  alpha1,
  beta1,
  x2,
  p2,
  q2,
  alpha2,
  beta2,
  theta,
  prec = 200
)
}
\arguments{
\item{x1}{first vector of relative abundances in the (x1, x2) pair.}

\item{p1}{zero-inflation probability for x1.}

\item{q1}{one-inflation probability for x1.}

\item{alpha1}{alpha parameter in beta distribution for x1.}

\item{beta1}{beta parameter in beta distribution for x1.}

\item{x2}{second vector of relative abundances in the (x1, x2) pair.}

\item{p2}{zero-inflation probability for x2.}

\item{q2}{one-inflation probability for x2.}

\item{alpha2}{alpha parameter in beta distribution for x2.}

\item{beta2}{beta parameter in beta distribution for x2.}

\item{theta}{copula dependence parameter.}
}
\value{
\code{bivariateDensityZOIB} returns the bivariate zero-one-inflated beta density function using a frank copula.
}
\description{
Bivariate frank copula density function with zero-inflated beta margins.
}
