% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copnet.R
\name{copnet}
\alias{copnet}
\title{copnet}
\usage{
copnet(abd, covars = NULL, ncores = 1)
}
\arguments{
\item{abd}{data frame of microbial relative abundances. Rows are samples and columns are taxa.}

\item{covars}{data frame of covariates to adjust for in the marginal regression models. Rows should be in the same order as \code{abd} and rownames should match.}

\item{ncores}{number of cores for parallelization. Default value is 1 (no parallel computing).}
}
\value{
\code{copnet} returns the two-stage estimate of \eqn{\theta} the LR statistics and the p-value.
}
\description{
Implementation of Copulas with Mixture Margins for Covariation Networks using two-stage estimation and testing.
}
