% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copnetLarge.R
\name{copnetLarge}
\alias{copnetLarge}
\title{Large Data get the MLE of marginal parameter and get the estimate}
\usage{
copnetLarge(abd, covars = NULL, ncores = 1, wd_Marginal, wd_estimate)
}
\arguments{
\item{abd}{data frame of microbial relative abundances. Rows are samples and columns are taxa.}

\item{covars}{data frame of covariates to adjust for in the marginal regression models. Rows should be in the same order as \code{abd} and rownames should match.}

\item{ncores}{number of cores for parallelization. Default value is 1 (no parallel computing).}

\item{wd_Marginal}{working directory file name for saving the rds file related to marginal MLE parameter}

\item{wd_estimate}{working directory file name for saving the rds file related to copula estimator and p-value}

\item{covar_formula_list}{list of formula.mu, formula.sigma, formula.nu, formula.tau}
}
\value{
\code{copnetLarge} returns (column numbers in the pair, estimated theta, LR statistic, p-value)
}
\description{
Get the MLE of marginal parameter for the Large Dataset and save it and then estimate the parameter
}
