% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzib.R
\name{dzib}
\alias{dzib}
\title{The Zero-One-Inflated Beta Distribution}
\usage{
dzib(x, p, q, alpha, beta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{p}{vector of zero-inflation probabilities.}

\item{q}{vector of one-inflation probabilities.}

\item{alpha}{vector of alpha in beta distribution}

\item{beta}{vector of beta in beta distribution}
}
\value{
\code{dzib} returns the density.
}
\description{
Density function for the zero-one-inflated beta distribution with zero-one-inflation probability equal to \code{p}, one-inflation probability equal to \code{q}, beta distribution alpha equal to \code{alpha} and beta distribution beta equal to \code{beta}.
}
