% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frank_theta_to_spearman.R
\name{frank_theta_to_spearman}
\alias{frank_theta_to_spearman}
\title{Convert theta in Frank copula function to spearman correlation}
\usage{
frank_theta_to_spearman(theta)
}
\arguments{
\item{theta}{The copula parameter in Frank copula function}
}
\value{
\code{frank_theta_to_spearman} returns the corresponding spearman correlation for copula parameter in Frank copula function
}
\description{
Convert theta in Frank copula function to spearman correlation
}
