% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankConditionalU.R
\name{frankConditionalU}
\alias{frankConditionalU}
\title{The Frank Conditional Copula}
\usage{
frankConditionalU(u, v, theta)
}
\arguments{
\item{u}{vector of uniform random variables.}

\item{v}{vector of uniform random variables.}

\item{theta}{copula dependence parameter.}
}
\value{
\code{frankConditionalU} returns the frank conditional copula distribution function, conditional on u.
}
\description{
Frank conditional copula function.
}
\examples{
frankConditionalU(u = 0.35, v = 0.5, theta = 2)

}
