% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankCopula.R
\name{frankCopula}
\alias{frankCopula}
\title{The Frank Copula}
\usage{
frankCopula(u, v, theta)
}
\arguments{
\item{u}{vector of uniform random variables.}

\item{v}{vector of uniform random variables.}

\item{theta}{copula dependence parameter.}
}
\value{
\code{frankCopula} returns the frank copula distribution function.
}
\description{
Frank copula function.
}
