% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankDensity.R
\name{frankDensity}
\alias{frankDensity}
\title{The Frank Copula Density}
\usage{
frankDensity(u, v, theta)
}
\arguments{
\item{u}{vector of uniform random variables.}

\item{v}{vector of uniform random variables.}

\item{theta}{copula dependence parameter.}
}
\value{
\code{frankDensity} returns the frank copula density function.
}
\description{
Frank copula density function.
}
