% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankLogLik.R
\name{frankLogLik}
\alias{frankLogLik}
\title{Log-likelihood for the Bivariate Zero-One-Inflated Beta Density}
\usage{
frankLogLik(theta, p1, q1, alpha1, beta1, p2, q2, alpha2, beta2, x)
}
\arguments{
\item{theta}{copula dependence parameter.}

\item{p1}{zero-inflation probability for x1.}

\item{q1}{one-inflation probability for x1.}

\item{alpha1}{alpha parameter in beta distribution for x1.}

\item{beta1}{beta parameter in beta distribution for x1.}

\item{p2}{zero-inflation probability for x2.}

\item{q2}{one-inflation probability for x2.}

\item{alpha2}{alpha parameter in beta distribution for x2.}

\item{beta2}{beta parameter in beta distribution for x2.}

\item{x}{data frame with two columns for bivariate pair (x1, x2).}
}
\description{
Log-likelihood for the bivariate frank copula density function with a zero-one-inflated beta margins.
}
