\name{gene_data}
\alias{gene_data}
\title{Gene Expression Example Data}
\description{
  A small dataset of sample gene methylation data and covariates related to cell for illustration.
}
\format{ A list with two datafram. `gen_data`: A data frame with 759 cell rows and 10 gene. `cov`: A data frame with 759 cell rows and 3 features.}
\usage{data(gene_data)}
