% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalMLE.R
\name{marginalMLE}
\alias{marginalMLE}
\title{Marginal Maximum Likelihood Estimation}
\usage{
marginalMLE(
  formula.mu = ~1,
  formula.sigma = ~1,
  formula.nu = ~1,
  formula.tau = ~1,
  df = NULL
)
}
\arguments{
\item{formula.mu}{an object of class formula: a symbolic description of the model to be fitted for the beta mean parameter.}

\item{formula.sigma}{an object of class formula: a symbolic description of the model to be fitted for the beta dispersion parameter.}

\item{formula.nu}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}

\item{formula.tau}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}

\item{df}{data frame of outcomes and covariates.(y, covariate1, covariate2, ...)}
}
\value{
\code{marginalMLE} returns a list of the following elements:
\item{coefficients}{vector of regression coefficients from the models for nu, tau, sigma, and mu.}
\item{fitted}{data frame of fitted values for nu, tau, sigma, and mu.}
}
\description{
Maximum likelihood estimation of marginal distribution function parameters.
}
