% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzib.R
\name{pzib}
\alias{pzib}
\title{The Zero-One-Inflated Beta Distribution}
\usage{
pzib(x, p, q, alpha, beta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{vector of alpha in beta distribution}

\item{beta}{vector of beta in beta distribution}

\item{nu}{vector of zero-inflation probabilities.}

\item{tau}{vector of one-inflation probabilities.}
}
\value{
\code{pzib} returns the distribution function.
}
\description{
Distribution function for the zero-one-inflated beta distribution with zero-one-inflation probability equal to \code{p}, one-inflation probability equal to \code{q}, beta distribution alpha equal to \code{alpha} and beta distribution beta equal to \code{beta}.
}
\examples{
pzib(x = 0.8, p = 0.2, q=0.3, alpha = 2, beta = 3)
}
