% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateBivariateZOIBRcov.R
\name{simulateBivariateZOIBRcov}
\alias{simulateBivariateZOIBRcov}
\title{Simulate the data pairs follows zero-one inflation data with frank copula association}
\usage{
simulateBivariateZOIBRcov(
  n,
  theta,
  formula.mu = ~1,
  formula.sigma = ~1,
  formula.nu = ~1,
  formula.tau = ~1,
  covariates = NULL,
  rho = NULL,
  delta = NULL,
  kappa = NULL,
  gamma = NULL
)
}
\arguments{
\item{n}{size of x data frame with two columns for bivariate pair (x1, x2).}

\item{theta}{copula dependence parameter.}

\item{formula.mu}{an object of class formula: a symbolic description of the model to be fitted for the beta mean parameter.}

\item{formula.sigma}{an object of class formula: a symbolic description of the model to be fitted for the beta dispersion parameter.}

\item{formula.nu}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}

\item{formula.tau}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}

\item{covariates}{data frame of covariate, Please include intercept as the first column}

\item{rho}{the list of the coefficient for nu rho[\link{1}] is the coefficient for x1 and  rho[\link{2}] is the coefficient for x2.}

\item{delta}{the list of the coefficient for mu delta[\link{1}] is the coefficient for x1 and  delta[\link{2}] is the coefficient for x2.}

\item{kappa}{the list of the coefficient for sigma kappa[\link{1}] is the coefficient for x1 and  kappa[\link{2}] is the coefficient for x2.}

\item{gamma}{the list of the coefficient for tau gamma[\link{1}] is the coefficient for x1 and  gamma[\link{2}] is the coefficient for x2.}
}
\description{
Simulate the data pairs follows zero-one inflation data with frank copula association
}
