% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaTSMLE.R
\name{thetaTSMLE}
\alias{thetaTSMLE}
\title{Two-stage Estimation of \eqn{\theta}}
\usage{
thetaTSMLE(x, lower, upper, p1, q1, alpha1, beta1, p2, q2, alpha2, beta2)
}
\arguments{
\item{x}{data frame with two columns for bivariate pair (x1, x2).}

\item{lower}{the lower end point of the interval to be searched in univariate optimizer.}

\item{upper}{the lower end point of the interval to be searched in univariate optimizer.}

\item{p1}{zero-inflation probability for x1.}

\item{q1}{one-inflation probability for x1.}

\item{alpha1}{alpha parameter in beta distribution for x1.}

\item{beta1}{beta parameter in beta distribution for x1.}

\item{p2}{zero-inflation probability for x2.}

\item{q2}{one-inflation probability for x2.}

\item{alpha2}{alpha parameter in beta distribution for x2.}

\item{beta2}{beta parameter in beta distribution for x2.}

\item{theta}{copula dependence parameter.}
}
\value{
\code{thetaTSMLE} returns the two-stage estimate of \eqn{\theta}.
}
\description{
Two-stage estimation of copula dependence parameter \eqn{\theta}.
}
