% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCoNet.R
\name{scCoNet}
\alias{scCoNet}
\title{Large Data get the MLE of marginal parameter and get the estimate}
\usage{
scCoNet(
  abd,
  covars = NULL,
  ncores = 1,
  sig_level = 0.05,
  sig_connect = FALSE,
  wd_Marginal = NULL,
  wd_estimate = NULL,
  formula.mu = y ~ 1,
  formula.sigma = ~1,
  formula.nu = ~1,
  formula.tau = ~1
)
}
\arguments{
\item{abd}{data frame of microbial relative abundances. Rows are samples and columns are taxa.}

\item{covars}{data frame of covariates to adjust for in the marginal regression models. Rows should be in the same order as \code{abd} and rownames should match.}

\item{ncores}{number of cores for parallelization. Default value is 1 (no parallel computing).}

\item{sig_level}{the significant level for select the connection based on the B-H adjusted p-value, defualt is 0.05.}

\item{sig_connect}{the argument for whether only return the significant connection or not. The default value is FALSE, it return all the connections.}

\item{wd_Marginal}{working directory file name for saving the rds file related to marginal MLE parameter, the default is NULL}

\item{wd_estimate}{working directory file name for saving the rds file related to copula estimator and p-value,, the default is NULL}

\item{formula.mu}{an object of class formula: a symbolic description of the model to be fitted for the beta mean parameter.}

\item{formula.sigma}{an object of class formula: a symbolic description of the model to be fitted for the beta dispersion parameter.}

\item{formula.nu}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}

\item{formula.tau}{an object of class formula: a symbolic description of the model to be fitted for the zero-one inflation probability parameter.}
}
\value{
\code{scCoNet} returns dataframe (Column 1 index, Column 2 index, estimated theta, transformed spearman, LR statistic, p-value, p-value after B-H adjustment)
}
\description{
Get the MLE of marginal parameter for the Large Dataset and save it and then estimate the parameter
}
